updates to FTP and main

+++ getFTP.py
@@ -1,5 +1,8 @@
 import subprocess
 
+#event TFI stuff
+# full pipeline
+
 def set_proxy(proxy):
     import os
     os.environ['http_proxy'] = proxy
@@ -10,7 +13,7 @@ def set_proxy(proxy):
     os.environ['FTP_PROXY'] = proxy
 
 
-def getFTPtar(date, s, dir):
+def getFTPtar(date, s, dir, tar=True):
     
 
     while True: #downloads and installs any missing packages
@@ -31,25 +34,39 @@ def getFTPtar(date, s, dir):
     # set_proxy("http://proxy-zsgov.external.lmco.com:80")
     #login to ftp server
     ftp = ftplib.FTP('ftp.swpc.noaa.gov')
-    print("here")
+    # print("here")
     ftp.login()
     ftp.cwd('pub/warehouse')
 
     #get a list of every folder that contains data we want
     files = []
-    print([file for file in ftp.nlst()])
+    # print([file for file in ftp.nlst()])
     for file in ftp.nlst(): #iterate through all files in /pub/warehouse
-        print(file)
-        # try: #the folders we want are named after the year in which the data was taken, we can therefore filter out any file that isn't a year number
-        #     int(file)
-        # except ValueError: #if the folder name is not a number, skip
-        #     continue
-        if type(file) == type(1):
-            if (int(file))*10000 >= d - d%10000: #make sure the folder year is after the specified date
+        # print(file, type(file))
+        try: #the folders we want are named after the year in which the data was taken, we can therefore filter out any file that isn't a year number
+            
+            int(file)
+            # print("here")
+            if (int(file))*10000 >= d - d%10000:
                 files.append(file)
+        except ValueError: #if the folder name is not a number, skip
+            continue
+        # if type(file) == type(1):
+        #     # if (int(file))*10000 >= d - d%10000: #make sure the folder year is after the specified date
+        #     print(file)
+        #     files.append(file)
     files.sort()
+    # files = files[-2:]
+    print(files, "hi")
 
     folder = "ftp_download_%s" % (series)
+
+
+    # MANIPULATE THIS PART TO SPEED UP SAVING PROCESS
+    # use this to see what the last text file in last year was
+    # only download the ones relevant if date given as input to function
+    # if not, just downlaod everything
+    # open existing tar if exists, add data, then recompress
     if os.path.exists(folder):
         while True: #this keeps breaking but it eventually works if you keep running it until it stops error-ing
             try:
@@ -57,16 +74,20 @@ def getFTPtar(date, s, dir):
                 break
             except OSError:
                 continue
+    
     os.mkdir(folder) #create directory for all files
     for year in files:
         file = "%s/%s_%s.tar.gz" % (year, year, series) #can't use new methods for string formatting since server might be running python 2
+        print(ftp.nlst(year))
         if file in ftp.nlst(year): #data for the current year isn't zipped, so there needs to be a different method for that 
+            print("here1")
             with open(year + ".tar.gz", 'wb') as fh: #retrieve data from ftp server
                 ftp.retrbinary("retr %s" % (file), fh.write)
             with tarfile.open(year + ".tar.gz") as fh: #unzip
                 fh.extractall(folder)
             os.remove(year + ".tar.gz") #delete zip file
         else:
+            print("hwerew2")
             f = "%s/%s_%s" % (folder, year, series) #path to folder for this year
             ftp_f = year + "/" + series
             if series == "events":
@@ -79,10 +100,15 @@ def getFTPtar(date, s, dir):
                     if int(file_.split('/')[-1][:8]) > d: ftp.retrbinary("retr %s" % (file_), fh.write)
                 d = int(file_.split('/')[-1][:8]) #update d
 
-    with tarfile.open(folder + ".tar.gz", "w:gz") as fh:
-        for root, dirs, files in os.walk(folder):
-            for file in files:
-                fh.add(os.path.join(root, file))
+    if tar:
+        with tarfile.open(folder + ".tar.gz", "w:gz") as fh:
+            for root, dirs, files in os.walk(folder):
+                for file in files:
+                    fh.add(os.path.join(root, file))
 
-    shutil.rmtree(folder) #remove all temporary files once zipped
+        shutil.rmtree(folder) #remove all temporary files once zipped
     print(d)
+getFTPtar(2022, "events", "/Users/justinhou/Documents/data")
+
+def extractFTP(series): # extract from tar
+    return True
